feat: new ConditionsExpression Enum move dynamo functions into dynamo folder (#178)
* feat: new ConditionsExpression Enum move dynamo functions into dynamo folder

* feat: moving Dynamo types into dynamo/types.ts
