+++ src/dao/index.ts
@@ -1,2 +1,3 @@
 export * from './interfaces'
 export * from './sqlUtils'
+export * from './dynamo'
