+++ src/dao/interfaces.ts
@@ -1,18 +1,3 @@
-export interface DynamodbParams {
-  TableName: any
-  Key?: any
-  Limit?: number
-  ExclusiveStartKey?: object
-  ExpressionAttributeNames?: object
-  ExpressionAttributeValues?: object
-  KeyConditionExpression?: string
-  Item?: object
-}
-
-export interface DynamodbResponseBatch {
-  UnprocessedItems: object
-}
-
 export interface SqlOperators {
   columnName: string
   value: string[] | string
